Version 1.0.3.
Version 1.0.3:
Fixed a bug causing the Monte Carlo procedure to produce exactly the same dataset

Version 1.0.2:
Fixed the duplicates algorithm so that compares the two variable names that have been formatted to 8 characters or fewer using the "format_mplus" function
Modified the "format_mplus" algorithm for formatting the Mplus variable names.
Fixed a bug in the observed correlation tables that would not allow you to display variable names / labels.

Version 1.0.1:
Fixed bug in Fit Statistics tabular output where the DF and p were not displayed for the chi-square statistic
Fixed a bug in which MEANS/INTERCEPTS/THRESHOLDS were not properly loaded from the sample statistics for categorical variables